iOS 9.0

+++ PrivateFrameworks/PerformanceAnalysis.framework/PAGenealogyTraceMessage.h
@@ -27,7 +27,7 @@
 - (id)binaryUuid;
 - (void)dealloc;
 - (id)initWithId:(unsigned long long)arg1 andOffset:(unsigned int)arg2 intoBinaryWithUUID:(id)arg3 withBuffer:(const void*)arg4 ofLength:(unsigned long)arg5 atMachTime:(double)arg6 withActivityId:(unsigned long long)arg7;
-- (id)initWithSerializedGenealogyTraceMessage:(const struct { unsigned long long x1; double x2; unsigned long long x3; unsigned long long x4; long long x5; unsigned int x6; unsigned int x7; }*)arg1;
+- (id)initWithSerializedGenealogyTraceMessage:(const struct { unsigned long long x1; double x2; unsigned long long x3; unsigned long long x4; long long x5; unsigned int x6; unsigned long long x7; }*)arg1;
 - (id)message;
 - (void)populateReferencesUsingBufferPosition:(const void*)arg1 andDeserializationDictionary:(id)arg2 andDataBufferDictionary:(id)arg3;
 - (unsigned long)sizeInBytesForSerializedVersion;
